Total overhaul of HTTP parsing.  Functionally equivalent but now set up with a more "streaming" (pull) structure instead of pushing bytes into a state machine.  Eventually this will enable us to dispatch "on the fly" while reading an HTTP message.  For example we could see if the user has any usage tokens before bothering to accept a file upload.
All sorts of possibilities arise when you can handle streaming.  When we do
this we'll break out of the read_complete_message paradigm and use
Loom::Sloop::HTTP::Parse as a tookit or event generator instead.  There's
already some of this going on *inside* the read_complete_message routine,
for example the calls to "read_line", "read_header", and "read_content".
But ultimately instead of "read_content" we'll want to parse the form
parameters and file uploads in pieces.  I'll keep factoring away on this
*internally* until the correct structure emerges, then I'll start migrating
it upwards to implement the streaming.

In light of this change, the Sloop architecture now expects the top level
application module (in this case Loom::Object::Loom::Main) to handle the
entire interaction with the client.  We now have a $client object with the
signature [receive, send, disconnect, exiting].

Even what we have already done is a major improvement in robustness and
efficiency.  We're using a circular byte buffer which is used and modified
"in place", and this minimizes copying of bytes coming in from the client
socket.  We also have more localized "reasonable limits" on parts of HTTP
message, for example limits on line length and limits on the number of
HTTP headers we will read.  This helps shield us from deliberately insane
incoming data.  We already had a limit on total message size but these new
localized limits tighten things down even more.

I also drastically simplified the "KV" (key-value) parser, getting rid of that
old bulky "Format" suite.

I renamed the "Site" directory as "Object", because that is more in line with
the coming ability to host arbitrary new object types in a Loom server.

I improved the handling of server config files using "on-demand" loading
(Loom::Sloop::Config.pm).

I abstracted the handling of signals (Loom::Sloop::Signal.pm), and also
the handling of global server state (Loom::Sloop::Status.pm).